+++ smarts/core/vehicle_index.py
@@ -601,7 +601,8 @@ class VehicleIndex:
         Vehicle.attach_sensors_to_vehicle(
             sim, vehicle, agent_interface, sensor_state.mission_planner
         )
-        if sim.renderer:
+        if sim.is_rendering:
+            vehicle.create_renderer_node(sim.renderer)
             sim.renderer.begin_rendering_vehicle(vehicle.id, is_agent=True)
 
         vehicle_id = _2id(vehicle.id)
@@ -636,7 +637,8 @@ class VehicleIndex:
         )
 
         vehicle_id, actor_id = _2id(vehicle_id), _2id(actor_id)
-        if sim.renderer:
+        if sim.is_rendering:
+            vehicle.create_renderer_node(sim.renderer)
             sim.renderer.begin_rendering_vehicle(vehicle.id, is_agent=False)
 
         self._vehicles[vehicle_id] = vehicle
@@ -653,6 +655,14 @@ class VehicleIndex:
         self._controlled_by = np.insert(self._controlled_by, 0, tuple(entity))
 
         return vehicle
+
+    def begin_rendering_vehicles(self, renderer):
+        agent_ids = self.agent_vehicle_ids()
+        for vehicle in self._vehicles.values():
+            if not vehicle.renderer:
+                vehicle.create_renderer_node(renderer)
+                is_agent = vehicle.id in agent_ids
+                renderer.begin_rendering_vehicle(vehicle.id, is_agent)
 
     def sensor_states_items(self):
         return map(lambda x: (self._2id_to_id[x[0]], x[1]), self._sensor_states.items())