+++ tools/pybullet_linear_tire_example.py
@@ -53,7 +53,7 @@ def look_at(client, position=(0, 0, 0), top_down=True):
         )
 
 
-def run(renderer, client, vehicle, plane_body_id, sliders, n_steps=1e6):
+def run(client, vehicle, plane_body_id, sliders, n_steps=1e6):
     prev_friction_sum = None
 
     controller_state = ActuatorDynamicControllerState()
@@ -167,8 +167,6 @@ def run(renderer, client, vehicle, plane_body_id, sliders, n_steps=1e6):
             break
 
         client.stepSimulation()
-        vehicle.sync_to_panda3d()
-        renderer.step()
 
         frictions_ = frictions(sliders)
 
@@ -209,9 +207,6 @@ def frictions(sliders):
 if __name__ == "__main__":
     # https://turtlemonvh.github.io/python-multiprocessing-and-corefoundation-libraries.html
     # mp.set_start_method('spawn', force=True)
-
-    # renderer = Renderer("linear_tier_example")
-    renderer = mock.MagicMock()
 
     client = bc.BulletClient(pybullet.GUI)
     # client = BulletClient(pybullet.GUI)
@@ -256,7 +251,6 @@ if __name__ == "__main__":
             vehicle = Vehicle(
                 "hello",
                 pose=pose,
-                renderer=renderer,
                 chassis=AckermannChassis(
                     pose=pose,
                     bullet_client=client,
@@ -265,7 +259,6 @@ if __name__ == "__main__":
             )
 
             run(
-                renderer,
                 client,
                 vehicle,
                 plane_body_id,